pyproject.toml
[tool.poetry]
name = "hackumass2024"
version = "0.1.0"
description = ""
authors = ["arfrie22 <43021241+arfrie22@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.9.0.80"
numpy = "^1.26.4"
matplotlib = "^3.9.0"
opencv-contrib-python = "^4.9.0.80"
scikit-image = "^0.23.2"
ipykernel = "^6.29.4"
torch = "^2.5.1"
torchvision = "^0.20.1"
scipy = "^1.14.1"
tqdm = "^4.67.0"
datasets = "^3.1.0"
transformers = {extras = ["torch"], version = "^4.46.2"}
accelerate = "^1.1.1"
scikit-learn = "^1.5.2"
pandas = "^2.2.3"
pyserial = "^3.5"
timm = "^1.0.11"
albumentations = "^1.4.21"
evaluate = "^0.4.3"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
